Switch from bower to npm for most dependencies
Prompted by @msecret to read the following:

http://gofore.com/ohjelmistokehitys/stop-using-bower/

This nudged me towards the convention of "rely on the npm by default" for
JavaScript dependencies. This necessitated bumping the Angular version to
1.3.x, as Angular wasn't properly browserify-able until 3.1.14:

https://github.com/angular/angular.js/blob/master/CHANGELOG.md#1314-instantaneous-browserification-2015-02-24

The new Angular version causes a ~19K bump in minified bundle size, ~7k
compressed:
  Before:
  133068 assets/js/search-bundle.js
   47680 assets/js/search-bundle.js.gz

  After:
  151892 assets/js/search-bundle.js
   54533 assets/js/search-bundle.js.gz

Bower is still used for angular-livesearch, but the debowerify transform is no
longer needed, as it just calls `angular.module()` and doesn't actually export
anything.