Remove dependency on angular-livesearch, part 1
This new implementation removes the dependency on angular-livesearch and
Angular.js in general. It's not yet complete; the UI elements for selecting
results need to be completely replaced. However, the mechanics of fetching and
loading the Lunr.js index and retrieving search results from the input box is
completely intact, illustrated by console messages.

The size of the new search bundle is now:

$ ls -l assets/js/
total 64
-rw-r--r--  1 ... ...  16479 Jan 28 18:46 search-bundle.js
-rw-r--r--  1 ... ...   5360 Jan 28 18:46 search-bundle.js.gz
-rw-r--r--  1 ... ...   2940 Jan 28 18:46 search.js

Compared to the previous implementation, this is a 9x reduction in the
uncompressed bundle size, 10x reduction in the compressed bundle size:

$ ls -l .../jekyll_pages_api_search-0.3.0/assets/js
total 424
-rw-r--r--  1 ... ...  151846 Aug 20 07:30 search-bundle.js
-rw-r--r--  1 ... ...   54514 Aug 20 07:30 search-bundle.js.gz
-rw-r--r--  1 ... ...    2842 Aug 20 07:30 search.js