Browserify renderJekyllPagesApiSearchResults
Permits users to define a script that exports a function conforming to the
renderJekyllPagesApiSearchResults function interface. Requires the user to
have installed browserify and uglifyify, but generates the bundle as part of
the normal `jekyll build` rather than as a separate step. Also injects a
`<script>` tag to load the custom bundle.